Edit test files to correctly reference component files
