+++ test/client/CallMeNow.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { shallow, mount, render } from 'enzyme';
+import { shallow } from 'enzyme';
 
-import { CallMeNow } from '../../client/src/components/CallMeNow.jsx';
+import { CallMeNow } from '../../client/src/components/entries/CallMeNow.jsx';
 
 describe('CallMeNow Component', () => {
