+++ test/client/EntryListItem.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { shallow, mount, render } from 'enzyme';
+import { mount } from 'enzyme';
 
-import Entry from '../../client/src/components/Entry.jsx';
+import Entry from '../../client/src/components/entries/EntryListItem.jsx';
 
 describe('Entry Component', () => {
   let entry;
