+++ test/client/Landing.test.js
@@ -1,17 +1,16 @@
 import React from 'react';
-import { shallow, mount, render } from 'enzyme';
+import { shallow } from 'enzyme';
 
-import { Home } from '../../client/src/components/Home.jsx';
+import { Home } from '../../client/src/components/landing/Index.jsx';
 
 describe('Home Component', () => {
   let home;
 
   beforeEach(() => {
     home = shallow(<Home />);
-  })
+  });
 
   it('should render without crashing', () => {
     expect(home.exists()).toBe(true);
   });
-
-})
+});
