+++ test/client/SignUpPhonePreferences.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import { shallow, mount, render } from 'enzyme';
+import { shallow } from 'enzyme';
 
-import { SignUpPhonePreferences } from '../../client/src/components/SignUpPhonePreferences.jsx';
+import { SignUpPhonePreferences } from '../../client/src/components/signup/PhonePreferences.jsx';
 
 describe('SignUpPhonePreferences Component', () => {
   let phonePreferences;
 
   beforeEach(() => {
     phonePreferences = shallow(<SignUpPhonePreferences />);
-  })
+  });
 
   it('should render without crashing', () => {
     expect(phonePreferences.exists()).toBe(true);
@@ -17,5 +17,4 @@ describe('SignUpPhonePreferences Component', () => {
   it('should render 3 select fields', () => {
     expect(phonePreferences.find('select').length).toEqual(3);
   });
-
-})
+});
