+++ test/client/SignUpPhoneVerification.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import { shallow, mount, render } from 'enzyme';
+import { shallow } from 'enzyme';
 
-import { SignUpPhoneVerification } from '../../client/src/components/SignUpPhoneVerification.jsx';
+import { SignUpPhoneVerification } from '../../client/src/components/signup/PhoneVerification.jsx';
 
 describe('SignUpPhoneVerification Component', () => {
   let phoneVerification;
 
   beforeEach(() => {
     phoneVerification = shallow(<SignUpPhoneVerification />);
-  })
+  });
 
   it('should render without crashing', () => {
     expect(phoneVerification.exists()).toBe(true);
@@ -17,5 +17,4 @@ describe('SignUpPhoneVerification Component', () => {
   it('should render 1 input field', () => {
     expect(phoneVerification.find('input').length).toEqual(1);
   });
-
-})
+});
